# docs/api/openapi.m13.json
{
...
  "components": {
    "schemas": {
...
      "ApprovalDecideDto": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "pattern": "^(approve|deny)$",
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ApprovalDecideDto",
        "type": "object"
      },
      "ApprovalDto": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "context": {
            "title": "Context",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "policy_source": {
            "title": "Policy Source",
            "type": "string"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "action",
          "risk",
          "context",
          "policy_source",
          "status",
          "version"
        ],
        "title": "ApprovalDto",
        "type": "object"
      },
      "BudgetViewDto": {
        "additionalProperties": false,
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/UsageEntryDto"
            },
            "title": "Entries",
            "type": "array"
          },
          "reservations": {
            "items": {
              "additionalProperties": false,
              "type": "object"
            },
            "title": "Reservations",
            "type": "array"
          },
          "total_estimated_cost_minor": {
            "title": "Total Estimated Cost Minor",
            "type": "integer"
          },
          "unknown_cost_entries": {
            "title": "Unknown Cost Entries",
            "type": "integer"
          }
        },
        "required": [
          "total_estimated_cost_minor",
          "unknown_cost_entries"
        ],
        "title": "BudgetViewDto",
        "type": "object"
      },
...
      "ClaimDto": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationDto"
            },
            "title": "Relations",
            "type": "array"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "statement",
          "status"
        ],
        "title": "ClaimDto",
        "type": "object"
      },
      "ClaimMapDto": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimDto"
            },
            "title": "Claims",
            "type": "array"
          },
          "contradictory_claims": {
            "items": {
              "type": "string"
            },
            "title": "Contradictory Claims",
            "type": "array"
          },
          "unsupported_claims": {
            "items": {
              "type": "string"
            },
            "title": "Unsupported Claims",
            "type": "array"
          }
        },
        "title": "ClaimMapDto",
        "type": "object"
      },
...
      "EvidenceDto": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "content_digest": {
            "title": "Content Digest",
            "type": "string"
          },
          "environment_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment Digest"
          },
          "experiment_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Run Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Digest"
          },
          "manifest_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Digest"
          },
          "model_refs": {
            "items": {
              "type": "string"
            },
            "title": "Model Refs",
            "type": "array"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_ref",
          "content_digest"
        ],
        "title": "EvidenceDto",
        "type": "object"
      },
      "ExportBundleDto": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimDto"
            },
            "title": "Claims",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDto"
            },
            "title": "Evidence",
            "type": "array"
          },
          "exported_from": {
            "title": "Exported From",
            "type": "string"
          },
          "manifest_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Digest"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_state": {
            "title": "Run State",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/BudgetViewDto"
          }
        },
        "required": [
          "run_id",
          "run_state",
          "usage",
          "exported_from"
        ],
        "title": "ExportBundleDto",
        "type": "object"
      },
      "HTTPValidationError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InterventionDto": {
        "additionalProperties": false,
        "description": "干预请求（M13：pause/resume 走状态机；语义变更 501）。",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "kind": {
            "pattern": "^(pause|resume|budget_adjust|replace_agent)$",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "InterventionDto",
        "type": "object"
      },
...
      "RelationDto": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "strength": {
            "default": 1.0,
            "title": "Strength",
            "type": "number"
          }
        },
        "required": [
          "claim_id",
          "evidence_id",
          "relation"
        ],
        "title": "RelationDto",
        "type": "object"
      },
...
      "UsageEntryDto": {
        "additionalProperties": false,
        "properties": {
          "actual_cost_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost Minor"
          },
          "cost_status": {
            "title": "Cost Status",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "estimated_cost_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Minor"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "entry_id",
          "resource_type",
          "quantity",
          "unit",
          "cost_status"
        ],
        "title": "UsageEntryDto",
        "type": "object"
      },
      "ValidationError": {
        "additionalProperties": false,
        "properties": {
          "ctx": {
            "additionalProperties": false,
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Research Console 控制面：配置 / 探测 / 运行 / 审批 / 审计",
    "title": "Research OS Control Plane API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/approvals/{approval_id}/decide": {
      "post": {
        "description": "后端裁决（hidden button != authorization）：\n\n- duplicate decide → 409；approve-vs-deny race → 后者 409；\n- stale If-Match → 412；run 非 WAITING_FOR_APPROVAL → 409；\n- deny → Run APPROVAL_REJECTED（FAILED）；approve → APPROVAL_GRANTED（RUNNING）；\n- 决策事件（approval.decided）落 outbox 审计。",
        "operationId": "decide_approvals__approval_id__decide_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecideDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/runs/{run_id}/claims": {
      "get": {
        "description": "Source → Evidence → Relation → Claim 地图（persisted truth）。\n\n视觉语义：missing/deleted evidence、contradiction（同一 claim 的\nSUPPORTS 与 REFUTES 并存）、unsupported claim（无 relation）在 DTO\n中显式表达；Claim 状态改变必须经过正式 use case/gate。",
        "operationId": "run_claim_map_runs__run_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimMapDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Claim Map",
        "tags": [
          "inspection"
        ]
      }
    },
...
    "/runs/{run_id}/evidence": {
      "get": {
        "description": "run 的 evidence（persisted truth；页面只 render）。",
        "operationId": "run_evidence_runs__run_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceDto"
                  },
                  "title": "Response Run Evidence Runs  Run Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evidence",
        "tags": [
          "inspection"
        ]
      }
    },
    "/runs/{run_id}/export": {
      "get": {
        "description": "Audit/Export：来自 persisted state（canonical 重算），不导出 UI 内存。",
        "operationId": "run_export_runs__run_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportBundleDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Export",
        "tags": [
          "inspection"
        ]
      }
    },
    "/runs/{run_id}/interventions": {
      "post": {
        "description": "干预：budget adjustment / future-agent replacement。\n\n运行中修改 Model/Tool/semantic 配置必须产生 Manifest Revision / Fork，\n不得 silent mutation；持久化配置修改属 M14。本端点对语义变更返回\n501（诚实边界），对纯状态类（pause/resume）走正式状态机。",
        "operationId": "intervene_runs__run_id__interventions_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Intervene Runs  Run Id  Interventions Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Intervene",
        "tags": [
          "approvals"
        ]
      }
    },
    "/runs/{run_id}/pause": {
      "post": {
        "description": "干预：pause（RUNNING → PAUSED；正式状态机迁移）。",
        "operationId": "pause_run_runs__run_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Pause Run Runs  Run Id  Pause Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Run",
        "tags": [
          "approvals"
        ]
      }
    },
    "/runs/{run_id}/resume": {
      "post": {
        "description": "干预：resume（PAUSED → RUNNING；正式状态机迁移）。",
        "operationId": "resume_run_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Resume Run Runs  Run Id  Resume Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Run",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/runs/{run_id}/usage": {
      "get": {
        "description": "Budget/Usage 视图：正式 UsageLedger truth。\n\n- UNKNOWN 成本显式标记（禁止显示 0）；\n- estimated/actual 区分（KNOWN 时 estimated_cost_minor 必有）；\n- reservations 来自正式 ledger snapshot。",
        "operationId": "run_usage_runs__run_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetViewDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Usage",
        "tags": [
          "inspection"
        ]
      }
    },
...
  }
}